Ex6 fully done
