+++ BitStream/V2/de_encoder_fc.c
@@ -1,5 +1,6 @@
 #include "de_encoder_fc.h"
 #include <stdio.h>
+#include <stdlib.h>
 //-----------------------
 //      GetStr 
 //-----------------------
@@ -9,7 +10,7 @@ char* getStr(){
     return a;
 }
 //-----------------------
-//      Encoding 
+//      Used in encoding 
 //-----------------------
 void writeBit(BitStream* bs,int bit){
     // idx inicial value 7
@@ -24,7 +25,7 @@ void writeBit(BitStream* bs,int bit){
 
 }
 //-----------------------
-//      Decoding 
+//      Used in decoding 
 //-----------------------
 int readBit(BitStream* bs){
 
@@ -66,4 +67,51 @@ void initBitStreams(BitStream* bsr, BitStream* bsw){
     bsw->buf = 0;
     bsw->fp = fopen("out_file","w");
 
+}
+//-----------------------
+//      Decoding - Done
+//-----------------------
+void decoder(BitStream* bsr_p,BitStream* bsw_p){
+
+    int on_off = readBit(bsr_p);  // Inicial value to tell the while whether to run the first iteration or not  
+
+    while( on_off != -1){        // cant have the function called here, because would be calling it twice
+
+        if(on_off){
+            printf("Adding bit 1 to file\n" );
+            putc(49,bsw_p->fp);
+        }
+        else{
+            printf("Adding bit 0 in file\n" );
+            putc(48,bsw_p->fp);
+        }
+        on_off = readBit(bsr_p);
+
+    }
+}
+//-----------------------
+//      Encoding  Src only '0' '1's
+//-----------------------
+void encoder(BitStream* bsr_p,BitStream* bsw_p){
+
+    // Inicialmente tenho de ir buscar o caracter '0' ou '1' que está no ficheiro de texto ( 0b00110001 ou 0b00110000)
+
+    int chr = getc(bsr_p->fp); // Inicial value to tell the while whether to run the first iteration or not
+    while( chr != EOF){
+        if(chr == 49){
+            printf("Inserted bit 1 into the selected byte\n");
+            writeBit(bsw_p,1);
+        }
+        else if(chr==48)
+        {
+            printf("Inserted bit 0 into the selected byte\n");
+            writeBit(bsw_p,0);
+        }else
+        {
+            long int ilegal_pos = ftell(bsr_p->fp);
+            printf("Unallowed character at position %ld\n.",ilegal_pos);
+            exit(0);
+        }
+            chr = getc(bsr_p->fp);   
+    }
 }
