+++ BitStream/V2/de_encoder_fc.h
@@ -15,10 +15,14 @@ void initBitStreams(BitStream* bsr, BitStream* bsw);
 
 void writeBit(BitStream* bs,int bit); // Function prototype
 
+void encoder(BitStream* bsr_p,BitStream* bsw_p);
+
 // A função writeBit guarda um bit na posição de memória e avança o cursor e o ponteiro do ficheiro, quando o cursor ultrapassa 8 escreve o byte no ficheiro, correspondendo a um qualquer caracter. É usado no encoding.
 // O argumento 0 ou 1 da função deve ser obtido pelo caracter representado no texto.
 
 int readBit(BitStream* bs);
+
+void decoder(BitStream* bsr_p,BitStream* bsw_p);
 
 //A função readBit guarda um byte numa posição de memória e avança a posição do ponteiro do ficheiro. Deve calcular no bit de memória onde o cursor está, devolvendo verdadeiro se o bit for 1 e falso se for 0, e em seguida incrementar o cursor. De acordo com o retorno da função,no main o programa deve escrever o valor equivalente a '1' ou '0' (ASCII) de modo a escrever os carateres em modo texto. É usado no decoding.
