+++ BitStream/V2/out_file
@@ -1,1 +1,1 @@
-10
+0011000100110000
