+++ BitStream/V2/src_file
@@ -1,1 +1,1 @@
-0011000100110000
+10
